moved common functions to common.c
