+++ iscos_tests/multi_stream_basic/normal.c
@@ -1,4 +1,4 @@
-#include "common.h"
+#include "common.c"
 
 /* RAND_MAX assumed to be 32767 */
 unsigned char myrand(void) {
@@ -23,7 +23,6 @@ void normal_write(void *x)
 		/* stream id is persistent in the kernel for an open fd.
 		   If a normal write is intented while at a stream is open, it
 		   is suggested to write a stream_id of 0 before the write */ 
-		posix_fadvise(fd, 0, 0, POSIX_FADV_STREAMID);
 		for (current_offset = IO_OFFSET_NW ; current_offset < MAX_FILE_OFFSET ; current_offset += IO_TRANSFER_SIZE) {
 			err = pwrite(fd, f->data_in, IO_TRANSFER_SIZE, current_offset); 
 			if (err<0) {
@@ -37,88 +36,6 @@ void normal_write(void *x)
 		printf("WRITE DONE: \n");
 	}
 	close(fd);
-}
-
-int nvme_dir_send(int fd, __u32 nsid, __u16 dspec, __u8 dtype, __u8 doper,
-                  __u32 data_len, __u32 dw12, void *data, __u32 *result)
-{
-        struct nvme_admin_cmd cmd = {
-                .opcode         = nvme_admin_directive_send,
-                .addr           = (__u64)(uintptr_t) data,
-                .data_len       = data_len,
-                .nsid           = nsid,
-                .cdw10          = data_len? (data_len >> 2) - 1 : 0,
-                .cdw11          = dspec << 16 | dtype << 8 | doper,
-                .cdw12          = dw12,
-        };
-        int err;
-
-        err = ioctl(fd, NVME_IOCTL_ADMIN_CMD, &cmd);
-        if (!err && result)
-                *result = cmd.result;
-        return err;
-}
-
-int nvme_dir_recv(int fd, __u32 nsid, __u16 dspec, __u8 dtype, __u8 doper,
-                  __u32 data_len, __u32 dw12, void *data, __u32 *result)
-{
-        struct nvme_admin_cmd cmd = {
-                .opcode         = nvme_admin_directive_recv,
-                .addr           = (__u64)(uintptr_t) data,
-                .data_len       = data_len,
-                .nsid           = nsid,
-                .cdw10          = data_len? (data_len >> 2) - 1 : 0,
-                .cdw11          = dspec << 16 | dtype << 8 | doper,
-                .cdw12          = dw12,
-        };
-        int err;
-
-        err = ioctl(fd, NVME_IOCTL_ADMIN_CMD, &cmd);
-        if (!err && result)
-                *result = cmd.result;
-        return err;
-}
-
-int nvme_get_nsid(int fd)
-{
-        static struct stat nvme_stat;
-        int err = fstat(fd, &nvme_stat);
-
-        if (err < 0)
-                return err;
-
-        if (!S_ISBLK(nvme_stat.st_mode)) {
-                fprintf(stderr,
-                        "Error: requesting namespace-id from non-block device\n");
-                exit(ENOTBLK);
-        }
-        return ioctl(fd, NVME_IOCTL_ID);
-}
-
-int enable_stream_directive(int fd)
-{
-	__u32 result;
-	int err;
-	int nsid = nvme_get_nsid(fd);
-
-	printf("Enable stream directive for nsid %d\n", nsid);
-	err = nvme_dir_send(fd, nsid, 0, 0, 1, 0, 0x101, NULL, &result);
-	// dspec, dtype, doper, length, dw12, data, result
-	// err = nvme_dir_send(fd, nsid, 0, 0, 4, 0, 0x101, NULL, &result);
-	return err;
-}
-
-int alloc_stream_resources(int fd, unsigned int rsc_cnt)
-{
-	__u32 result;
-	int err;
-	int nsid = nvme_get_nsid(fd);
-
-	printf("Allocate stream resource for nsid %d\n", nsid);
-	err = nvme_dir_recv(fd, nsid, 0, 1, 3, 0, rsc_cnt, NULL, &result);
-	if (err==0)
-		printf("  requested %d; returned %d\n", rsc_cnt, result & 0xffff);
-	return err;
 }
 
 int main(int argc, char **argv)
@@ -167,7 +84,6 @@ int main(int argc, char **argv)
 
 	for (i=0; i<IO_SEGMENT_SIZE; i++)
 		f.data_in[i] = (unsigned char)(i % 100);
-	f.sid = 1;
 
 	normal_write(&f);
