+++ iscos_tests/multi_stream_basic/stream.c
@@ -1,4 +1,4 @@
-#include "common.h"
+#include "common.c"
 
 /* RAND_MAX assumed to be 32767 */
 unsigned char myrand(void) {
@@ -19,7 +19,6 @@ void stream_write(void *x)
 	else {
 		int err;
 		printf("Opened fd from st_thread: %d\n", fd);
-		posix_fadvise(fd, f->sid, 0, POSIX_FADV_STREAMID);
 		err = pwrite(fd, f->data_in, IO_ST_TRANSFER_SIZE, IO_OFFSET_ST); 
 		if (err<0)
 			fprintf(stderr, "nvme write from st_thread status:%#x(%s) \n", errno, strerror(errno));
@@ -32,86 +31,6 @@ void stream_write(void *x)
 	close(fd);
 	return;
 
-}
-
-int nvme_dir_send(int fd, __u32 nsid, __u16 dspec, __u8 dtype, __u8 doper,
-                  __u32 data_len, __u32 dw12, void *data, __u32 *result)
-{
-        struct nvme_admin_cmd cmd = {
-                .opcode         = nvme_admin_directive_send,
-                .addr           = (__u64)(uintptr_t) data,
-                .data_len       = data_len,
-                .nsid           = nsid,
-                .cdw10          = data_len? (data_len >> 2) - 1 : 0,
-                .cdw11          = dspec << 16 | dtype << 8 | doper,
-                .cdw12          = dw12,
-        };
-        int err;
-
-        err = ioctl(fd, NVME_IOCTL_ADMIN_CMD, &cmd);
-        if (!err && result)
-                *result = cmd.result;
-        return err;
-}
-
-int nvme_dir_recv(int fd, __u32 nsid, __u16 dspec, __u8 dtype, __u8 doper,
-                  __u32 data_len, __u32 dw12, void *data, __u32 *result)
-{
-        struct nvme_admin_cmd cmd = {
-                .opcode         = nvme_admin_directive_recv,
-                .addr           = (__u64)(uintptr_t) data,
-                .data_len       = data_len,
-                .nsid           = nsid,
-                .cdw10          = data_len? (data_len >> 2) - 1 : 0,
-                .cdw11          = dspec << 16 | dtype << 8 | doper,
-                .cdw12          = dw12,
-        };
-        int err;
-
-        err = ioctl(fd, NVME_IOCTL_ADMIN_CMD, &cmd);
-        if (!err && result)
-                *result = cmd.result;
-        return err;
-}
-
-int nvme_get_nsid(int fd)
-{
-        static struct stat nvme_stat;
-        int err = fstat(fd, &nvme_stat);
-
-        if (err < 0)
-                return err;
-
-        if (!S_ISBLK(nvme_stat.st_mode)) {
-                fprintf(stderr,
-                        "Error: requesting namespace-id from non-block device\n");
-                exit(ENOTBLK);
-        }
-        return ioctl(fd, NVME_IOCTL_ID);
-}
-
-int enable_stream_directive(int fd)
-{
-	__u32 result;
-	int err;
-	int nsid = nvme_get_nsid(fd);
-
-	printf("Enable stream directive for nsid %d\n", nsid);
-	err = nvme_dir_send(fd, nsid, 0, 0, 1, 0, 0x101, NULL, &result);
-	return err;
-}
-
-int alloc_stream_resources(int fd, unsigned int rsc_cnt)
-{
-	__u32 result;
-	int err;
-	int nsid = nvme_get_nsid(fd);
-
-	printf("Allocate stream resource for nsid %d\n", nsid);
-	err = nvme_dir_recv(fd, nsid, 0, 1, 3, 0, rsc_cnt, NULL, &result);
-	if (err==0)
-		printf("  requested %d; returned %d\n", rsc_cnt, result & 0xffff);
-	return err;
 }
 
 int main(int argc, char **argv)
@@ -160,7 +79,7 @@ int main(int argc, char **argv)
 
 	for (i=0; i<IO_SEGMENT_SIZE; i++)
 		f.data_in[i] = myrand();
-	f.sid = 1;
+	//f.stream_id = 1;
 
 	stream_write(&f);
